TensorFlow to develop a digital twin
The digital twinning is scheduled to be demonstrated on the design model developed on the ANN/CNN at the cloud data clusters
Here is a Python program to be enabled on TensorFlow to develop a digital twin to be demonstrated on the design model developed on the ANN/CNN at the cloud data clusters. It is set, designed, developed, trained, and to be updated at Google Cloud using the TensorFlow. The digital twin is replicated against physical surrogates built by physical data and data browsed at the internet using datasets at the Cloud, all through the product life cycle: